fix: relocate About out of src/pages to stop Pages Router duplication
Having any file under src/pages activates Next's Pages Router alongside
the App Router. It was serving /About (Pages) in parallel with /about
(App), producing duplicate content with separate bundles.

- Move src/pages/About.tsx -> src/components/About.tsx
- Update import in src/app/about/page.tsx
- Drop the now-dead src/pages glob from tailwind.config.js
