+++ tailwind.config.js
@@ -3,7 +3,6 @@ export default {
   content: [
     './src/app/**/*.{ts,tsx}',
     './src/components/**/*.{ts,tsx}',
-    './src/pages/**/*.{ts,tsx}',
     './src/content/**/*.{ts,tsx}',
   ],
   darkMode: 'class',
